+++ tests/test_data.py
@@ -23,7 +23,6 @@ def test_halos_unclean(tmp_path):
 
     # to regenerate reference
     #ref = cat.halos
-    #import asdf; asdf.compression.set_compression_options(typesize='auto')
     #ref.write(HALOS_OUTPUT_UNCLEAN, all_array_storage='internal', all_array_compression='blsc')
 
     ref = Table.read(HALOS_OUTPUT_UNCLEAN)
@@ -47,7 +46,6 @@ def test_halos_clean(tmp_path):
 
     # to regenerate reference
     #ref = cat.halos
-    #import asdf; asdf.compression.set_compression_options(typesize='auto')
     #ref.write(HALOS_OUTPUT_CLEAN, all_array_storage='internal', all_array_compression='blsc')
 
     ref = Table.read(HALOS_OUTPUT_CLEAN)
@@ -73,8 +71,19 @@ def test_subsamples_unclean(tmp_path):
     '''
 
     from abacusnbody.data.compaso_halo_catalog import CompaSOHaloCatalog
+    
+    cat = CompaSOHaloCatalog(os.path.join(EXAMPLE_SIM, 'halos', 'z0.000'), subsamples=dict(A=True), fields='all', cleaned=False)
+    lenA = len(cat.subsamples)
+    assert lenA == 2975
+    assert cat.subsamples.colnames == ['pos', 'vel']
+    
+    cat = CompaSOHaloCatalog(os.path.join(EXAMPLE_SIM, 'halos', 'z0.000'), subsamples=dict(B=True), fields='all', cleaned=False)
+    lenB = len(cat.subsamples)
+    assert lenB == 7082
 
     cat = CompaSOHaloCatalog(os.path.join(EXAMPLE_SIM, 'halos', 'z0.000'), subsamples=True, fields='all', cleaned=False)
+    
+    assert len(cat.subsamples) == lenA + lenB
 
     # to regenerate reference
     #ref = cat.subsamples
@@ -175,3 +184,17 @@ def test_unpack_bits():
     # bad bits field name
     with pytest.raises(ValueError):
         cat = CompaSOHaloCatalog(os.path.join(EXAMPLE_SIM, 'halos', 'z0.000'), subsamples=True, unpack_bits=['blah'], fields='N')
+
+
+def test_filter_func():
+    '''Test CHC filter_func
+    '''
+    
+    from abacusnbody.data.compaso_halo_catalog import CompaSOHaloCatalog
+
+    cat = CompaSOHaloCatalog(os.path.join(EXAMPLE_SIM, 'halos', 'z0.000'), fields=['N','x_L2com'],
+                            filter_func = lambda c: c['N'] > 100,
+                            subsamples=True)
+    assert (cat.halos['N'] > 100).all()
+    assert len(cat.halos) == 146
+    assert len(cat.subsamples) == 7193